ingestion/requirements.txt
pdfplumber>=0.10.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
openai>=1.35.0
tiktoken>=0.7.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0
